+++ js/get-descriptions.js
@@ -1,0 +1,29 @@
+import {getRandomElement} from './util.js';
+
+const descriptions = [
+  'Ахаха, зацени, какой у него маленький объектив!',
+  'Потенциальной добычи так много, что можно никуда не торопиться и немного потанцевать.',
+  'Жизнь грустная, зато зарплата смешная',
+  'Давайте рассмотрим изображение внимательнее',
+  'Я считаю, что снимок получился завораживающим',
+  'Ничего личного, просто бизнес.',
+  'Красиво жить не запретишь.',
+  'Только мертвая рыба плывет по течению.',
+  'Жизнь коротка. Улыбайтесь, пока у вас еще есть зубы.',
+  'Фотография жены в моём бумажнике, напоминает мне о том, что на этом месте могли бы быть деньги…',
+  'Невозможно жить полной жизнью на пустой желудок!',
+  'Не так страшны корпоративы, как «вас отметили на фото».',
+  'Эта жара заставила меня понять, что я не выживу в аду. Пора меняться…',
+  'Хорошо там, где меня нет… Но ничего, я и туда доберусь!',
+  'Не смотри на надпись, смотри на меня',
+  'Профессиональная выдержка',
+  'Иногда, когда я закрываю глаза, становится темно.',
+  'Я почти уверен, что вы смотрите на мое изображение',
+  'Смотреть можно, но трогать нельзя',
+];
+
+function getDescriptions() {
+  return getRandomElement(descriptions);
+}
+
+export {getDescriptions};
